fixed some bugs

+++ src/main/java/com/stockbroker/model/StockLot.java
@@ -5,36 +5,36 @@ import lombok.Getter;
 @Getter
 public class StockLot {
     private final Stock stock;
-    private int quantity;
+    private long quantity;
     private double avgBuyPrice;
 
-    public StockLot(Stock stock, int quantity) {
+    public StockLot(Stock stock, long quantity) {
         this.stock = stock;
         this.quantity = quantity;
         this.avgBuyPrice = quantity > 0 ? stock.getCurrPrice() : 0;
     }
 
-    public void addQuantity(int quantity) {
+    public void addQuantity(long quantity) {
         double oldPrice = this.avgBuyPrice;
-        int oldQuantity = this.quantity;
+        long oldQuantity = this.quantity;
         double newPrice = stock.getCurrPrice();
-        int newQuantity = this.quantity + quantity;
+        long newQuantity = this.quantity + quantity;
         this.avgBuyPrice = getAveragePrice(oldPrice, oldQuantity, newPrice, newQuantity);
         this.quantity = newQuantity;
     }
 
-    public void removeQuantity(int quantity) {
+    public void removeQuantity(long quantity) {
         if (this.quantity < quantity)
             throw new RuntimeException("invalid operation due to insufficient quantity");
         this.quantity -= quantity;
     }
 
-    private double getAveragePrice(double oldPrice, int oldQuantity, double newPrice, int newQuantity) {
+    private double getAveragePrice(double oldPrice, long oldQuantity, double newPrice, long newQuantity) {
         if (newQuantity == 0)
             return 0;
         double oldAmount = oldPrice * oldQuantity;
         double newAmount = newPrice * newQuantity;
-        int totalQuantity = newQuantity + oldQuantity;
+        long totalQuantity = newQuantity + oldQuantity;
         double totalAmount = newAmount + oldAmount;
         return (totalAmount / totalQuantity);
