+++ src/main/java/com/stockbroker/model/stockexchange/ExchangeOrder.java
@@ -1,7 +1,6 @@
 package com.stockbroker.model.stockexchange;
 
 import com.stockbroker.model.OrderType;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,14 +8,14 @@ import lombok.Setter;
 @Setter
 public class ExchangeOrder {
     private int id;
-    private int quantity;
+    private long quantity;
     private String dematAccountId;
     private String symbol;
     private String stockGatewayName;
     private OrderType type;
     boolean isInitialOrder;
 
-    public ExchangeOrder(int quantity, String dematAccountId, String symbol, String stockGatewayName, OrderType type) {
+    public ExchangeOrder(long quantity, String dematAccountId, String symbol, String stockGatewayName, OrderType type) {
         this.quantity = quantity;
         this.dematAccountId = dematAccountId;
         this.symbol = symbol;
